split modules into submodules
